backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
aiosqlite>=0.19.0
aiohttp>=3.9.0
pandas>=2.0.0
numpy>=1.24.0
pypfopt>=1.5.5
python-dotenv>=1.0.0
pydantic>=2.0.0
